backend/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
requests==2.31.0
python-dotenv==1.0.0
psycopg[binary]==3.1.18
pytest==7.4.2
pytest-flask==1.2.0
gunicorn==21.2.0
redis==4.6.0
